# contracts/control-plane-v1/control-plane.openapi.yaml
openapi: 3.1.0
info:
  title: AI Agent Desktop Local Control Plane API
  version: 1.0.0
  summary: Stable local management contract for the desktop GUI and trusted clients.
  description: |
    The read-only foundation, cc-connect Windows AMD64 isolated installation,
    product-owned non-secret configuration revisions, explicit lifecycle
    ownership, and managed process lifecycle are implemented. Real credential
    writes, Telegram binding, other component lifecycle, and GUI remain
    unsupported. Provider-specific structures remain behind Adapter boundaries.
  license:
    name: See repository LICENSE
    identifier: Apache-2.0
  x-contract-status: frozen
  x-implementation-status: cc-connect-managed-runtime-partially-verified
  x-event-types:
    - com.aiagentdesktop.component.state.changed.v1
    - com.aiagentdesktop.component.discovered.v1
    - com.aiagentdesktop.provider.state.changed.v1
    - com.aiagentdesktop.registry.changed.v1
    - com.aiagentdesktop.operation.started.v1
    - com.aiagentdesktop.operation.progress.v1
    - com.aiagentdesktop.operation.completed.v1
    - com.aiagentdesktop.operation.failed.v1
    - com.aiagentdesktop.operation.cancel_requested.v1
    - com.aiagentdesktop.scan.progress.v1
    - com.aiagentdesktop.plan.generated.v1
    - com.aiagentdesktop.diagnostic.created.v1
    - com.aiagentdesktop.task.state.changed.v1
    - com.aiagentdesktop.log.appended.v1
servers:
  - url: http://127.0.0.1:{port}/api/v1
...
    variables:
      port:
...
security:
  - localBearer: []
...
paths:
...
  "/components/{componentId}/configuration-plans":
    parameters:
      - $ref: "#/components/parameters/ComponentId"
    post:
      tags: [Configuration]
      operationId: createManagedConfigurationPlan
      summary: Create an immutable non-secret product configuration plan.
      x-stability: stable
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                listen_port: { type: [integer, "null"], minimum: 59000, maximum: 59999 }
                expires_in_seconds: { type: integer, minimum: 60, maximum: 3600, default: 900 }
                rollback_to_revision: { type: [integer, "null"], minimum: 1 }
      responses:
        "201":
          description: Persisted plan waiting for explicit confirmation.
          content:
            application/json:
              schema:
                $ref: "./managed-runtime.schema.json#/$defs/ConfigurationPlan"
        default:
          $ref: "#/components/responses/Problem"
  "/components/{componentId}/configuration-plans/{planId}":
    parameters:
      - $ref: "#/components/parameters/ComponentId"
      - name: planId
        in: path
        required: true
        schema: { $ref: "#/components/schemas/Id" }
    get:
      tags: [Configuration]
      operationId: getManagedConfigurationPlan
      summary: Read one immutable managed configuration plan.
      responses:
        "200":
          description: Managed configuration plan.
          content:
            application/json:
              schema:
                $ref: "./managed-runtime.schema.json#/$defs/ConfigurationPlan"
        default:
          $ref: "#/components/responses/Problem"
  "/components/{componentId}/configuration:apply":
    parameters:
      - $ref: "#/components/parameters/ComponentId"
    post:
      tags: [Configuration]
      operationId: applyManagedConfigurationPlan
      summary: Confirm and atomically apply an immutable configuration plan.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [plan_id, plan_digest, current_revision, target_revision, confirmation]
              properties:
                plan_id: { type: string }
                plan_digest: { type: string, pattern: "^sha256:[a-f0-9]{64}$" }
                current_revision: { type: integer, minimum: 0 }
                target_revision: { type: integer, minimum: 1 }
                confirmation: { const: true }
      responses:
        "202": { $ref: "#/components/responses/OperationAccepted" }
        default: { $ref: "#/components/responses/Problem" }
...
  "/components/{componentId}/configuration":
    parameters:
      - $ref: "#/components/parameters/ComponentId"
    get:
      tags: [Configuration]
      operationId: getComponentConfiguration
      summary: Read a redacted configuration and its owner/revision.
      x-stability: stable
      parameters:
        - $ref: "#/components/parameters/ScopeIdQuery"
      responses:
        "200":
          description: Redacted configuration. Secret values are never returned.
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
          content:
            application/json:
              schema:
                $ref: "./core-models.schema.json#/$defs/RedactedConfig"
        default:
          $ref: "#/components/responses/Problem"
    put:
      tags: [Configuration]
      operationId: putComponentConfiguration
      summary: Atomically write a non-secret configuration for the current owner.
      x-stability: stable
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/IfMatch"
        - $ref: "#/components/parameters/CorrelationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfigurationWriteRequest"
      responses:
        "202":
          $ref: "#/components/responses/OperationAccepted"
        "409":
          $ref: "#/components/responses/Problem"
        "412":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"
  "/components/{componentId}/configuration:validate":
    parameters:
      - $ref: "#/components/parameters/ComponentId"
    post:
      tags: [Configuration]
      operationId: validateComponentConfiguration
      summary: Validate current or candidate configuration without activating it.
      x-stability: stable
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/CorrelationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfigurationValidationRequest"
      responses:
        "202":
          $ref: "#/components/responses/OperationAccepted"
        default:
          $ref: "#/components/responses/Problem"
...
  "/operations/{operationId}":
    parameters:
      - $ref: "#/components/parameters/OperationId"
    get:
      tags: [Operations]
      operationId: getOperation
      summary: Read current progress and the mutually exclusive result or error.
      x-stability: stable
      responses:
        "200":
          description: Operation snapshot.
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
          content:
            application/json:
              schema:
                $ref: "./core-models.schema.json#/$defs/Operation"
        "404":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"
...
  /external-tools/cc-switch:
    get:
      tags: [Maintenance]
      operationId: getCcSwitchExternalToolStatus
      summary: Detect CC Switch and disclose only verified external capabilities.
      responses:
        "200":
          description: CC Switch detection and capability evidence.
          content:
            application/json:
              schema: { $ref: "./managed-runtime.schema.json#/$defs/ExternalToolStatus" }
        default: { $ref: "#/components/responses/Problem" }
  "/external-tools/cc-switch:launch":
    post:
      tags: [Maintenance]
      operationId: launchCcSwitchExternalTool
      summary: Open a detected CC Switch executable without GUI automation or configuration writes.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [confirmation]
              properties:
                confirmation: { const: true }
      responses:
        "200":
          description: Normal launch receipt; it never claims install, update, or configuration success.
          content:
            application/json:
              schema: { type: object }
        default: { $ref: "#/components/responses/Problem" }
  /components/cc-connect/update-assessment:
    get:
      tags: [Maintenance]
      operationId: assessLockedCcConnectUpdate
      summary: Assess an exact locked target without performing an update.
      parameters:
        - name: requested_version
          in: query
          required: true
          schema: { type: string, minLength: 1, maxLength: 128 }
      responses:
        "200":
          description: Read-only update assessment.
          content:
            application/json:
              schema: { $ref: "./managed-runtime.schema.json#/$defs/UpdateAssessment" }
        default: { $ref: "#/components/responses/Problem" }
  /components/hermes/update-assessment:
    get:
      tags: [Maintenance]
      operationId: assessHermesUpdateBoundary
      summary: Return unsupported until a verified Hermes update Adapter exists.
      parameters:
        - name: requested_version
          in: query
          required: true
          schema: { type: string, minLength: 1, maxLength: 128 }
      responses:
        "200":
          description: Explicit unsupported assessment; no Hermes action occurs.
          content:
            application/json:
              schema: { $ref: "./managed-runtime.schema.json#/$defs/UpdateAssessment" }
        default: { $ref: "#/components/responses/Problem" }
...
components:
  securitySchemes:
    localBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque-local-token
      description: At least 256 bits of entropy. Supplied only in the Authorization header.
  headers:
    ETag:
      description: Opaque resource revision for optimistic concurrency.
      schema:
        type: string
...
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Reusing a key with a different normalized request returns IDEMPOTENCY_KEY_REUSE.
      schema:
        type: string
        minLength: 16
        maxLength: 256
    CorrelationId:
      name: X-Correlation-ID
      in: header
      required: false
      schema:
        $ref: "#/components/schemas/Id"
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: Opaque expected revision. A stale value returns REVISION_CONFLICT.
      schema:
        type: string
        minLength: 1
        maxLength: 128
    ComponentId:
      name: componentId
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Id"
...
    OperationId:
      name: operationId
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Id"
    ScopeIdQuery:
      name: scope_id
      in: query
      required: true
      schema:
        $ref: "#/components/schemas/Id"
...
  responses:
    OperationAccepted:
      description: Operation accepted. This does not prove that an external side effect has completed.
      headers:
        Location:
          description: Relative Operation resource URL.
          schema:
            type: string
        Retry-After:
          description: Suggested minimum polling interval in seconds when SSE is unavailable.
          schema:
            type: integer
            minimum: 1
      content:
        application/json:
          schema:
            $ref: "./core-models.schema.json#/$defs/Operation"
    Problem:
      description: RFC 9457 compatible, user-facing and redacted problem details.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
    Id:
      type: string
      minLength: 1
      maxLength: 128
      pattern: "^[A-Za-z0-9][A-Za-z0-9._:-]*$"
...
    ConfigurationValidationRequest:
      type: object
      additionalProperties: false
      required: [scope_id, mode, connectivity_check]
      properties:
        scope_id:
          $ref: "#/components/schemas/Id"
        mode:
          type: string
          enum: [current, candidate]
        candidate:
          type: object
          additionalProperties: true
        expected_revision:
          type: string
          maxLength: 128
        connectivity_check:
          type: boolean
          default: false
        deadline_at:
          type: string
          format: date-time
      allOf:
        - if:
            properties:
              mode:
                const: candidate
          then:
            required: [candidate, expected_revision]
            properties:
              candidate:
                type: object
                additionalProperties: true
              expected_revision:
                type: string
                maxLength: 128
    ConfigurationWriteRequest:
      type: object
      additionalProperties: false
      required: [scope_id, values, owner_type, expected_revision]
      properties:
        scope_id:
          $ref: "#/components/schemas/Id"
        values:
          type: object
          additionalProperties: true
          description: Non-secret values and credential references only.
        owner_type:
          type: string
          enum: [application, official_login, cc_switch, external]
        expected_revision:
          type: string
          minLength: 1
          maxLength: 128
        deadline_at:
          type: string
          format: date-time
...
    Problem:
      type: object
      additionalProperties: false
      required: [type, title, status, code, detail, user_message, retryable, recovery_actions, correlation_id]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
          minLength: 1
          maxLength: 256
        status:
          type: integer
          minimum: 400
          maximum: 599
        code:
          type: string
          pattern: "^[A-Z][A-Z0-9_]{2,127}$"
        detail:
          type: string
          maxLength: 2048
        user_message:
          type: string
          minLength: 1
          maxLength: 2048
        retryable:
          type: boolean
        recovery_actions:
          type: array
          uniqueItems: true
          items:
            type: string
            minLength: 1
            maxLength: 128
        diagnostic_id:
          $ref: "#/components/schemas/Id"
        operation_id:
          $ref: "#/components/schemas/Id"
        correlation_id:
          $ref: "#/components/schemas/Id"